moved thumbnails functionality to the correct file
