+++ client/components/Overview/Overview.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 const axios = require('axios');
 
 import ProductInformation from './Product-information.jsx';
-import SelectStyle from './Style-selector.jsx';
+import SelectStyle from './Style-thumbnails.jsx';
 import ShareOnline from './Share-to-socials.jsx';
 import AddToBag from './Add-to-bag.jsx';
 
@@ -11,16 +11,14 @@ const Overview = (props) => {
   // console.log('props', props);
   // move all the states into here
   // for styles
-  const [mainPhoto, setMainPhoto] = useState(props.currentStyle.results[0].photos[3].url);
+  const [mainPhoto, setMainPhoto] = useState(props.currentStyle.results[0].photos[0].url);
   const [styleName, setStyleName] = useState(props.currentStyle.results[0].name);
   // for size and quantity
   const [sizeAndStock, setSizeAndStock] = useState({});
 
   // event handler for clicking new image
   const handleThumb = (item) => {
-    console.log('CLICKED', item);
     setMainPhoto(item.url);
-    // console.log('TARGET?', mainPhoto);
   }
 
   return (
